% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qcbsmdc_get_files.R
\name{qcbsmdc_get_files}
\alias{qcbsmdc_get_files}
\title{Save files from a given table from the QCBS Mobile Data Collection System to your computer}
\usage{
qcbsmdc_get_files(table, folder = getwd())
}
\arguments{
\item{table}{a connection list file containing user_id and api_key values}

\item{folder}{the full path to the folder where you want the files to be saved}
}
\value{
saves the files to local directory
}
\description{
Save files from a given table from the QCBS Mobile Data Collection System to your computer
}
\examples{
qcbsmdc_get_files('TablePhotos','C:/User/MyName/MyFiles')
}
\seealso{
\code{\link{qcbsmdc_list_forms}} for obtaining forms

\code{\link{qcbsmdc_get_tables}} for obtaining the content of the tables

\code{\link{qcbsmdc_list_tables}} for obtaining the content of the tables
}

