% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qcbsmdc_get_tables.R
\name{qcbsmdc_get_tables}
\alias{qcbsmdc_get_tables}
\title{Get the content of all tables for a given form, or of a particular table, on the QCBS Mobile Data Collection System}
\usage{
qcbsmdc_get_tables(conn, form_id, table = "all")
}
\arguments{
\item{conn}{a connection list file containing user_id and api_key values}

\item{form_id}{the name or id of the form from which the table list is to be obtained}

\item{table}{either 'all' if all tables are desired, or the name of the table to be obtained}
}
\value{
list containing all tables associated with the form as data frames (if table='all) or data frame if the table option specifies the name of a particular table
}
\description{
Get the content of all tables for a given form, or of a particular table, on the QCBS Mobile Data Collection System
}
\examples{
## api keys can be obtained from your account on the http://quebio.ca/odk website
conn=list(user_id="JohnDoe",api_key='8fjasdf3h112aad')
#GET all tables
table_list<-qcbsmdc_get_tables(conn=conn,form_id='MyForm1',table='all')
#GET a particular table
the_table<-qcbsmdc_get_tables(conn=conn,form_id='MyForm1',table='uuid0MyTable')
}
\seealso{
\code{\link{qcbsmdc_list_forms}} for obtaining forms

\code{\link{qcbsmdc_list_tables}} for obtaining the content of the tables

\code{\link{qcbsmdc_get_files}} for saving files (e.g. photos) to your a local directory
}

