% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcbsmdc_list_forms.R
\name{qcbsmdc_list_forms}
\alias{qcbsmdc_list_forms}
\title{Get the list of forms for a given user on the QCBS Mobile Data Collection System}
\usage{
qcbsmdc_list_forms(conn)
}
\arguments{
\item{conn}{a connection list file containing user_id and api_key values}
}
\value{
character vector with list of form names
}
\description{
Get the list of forms for a given user on the QCBS Mobile Data Collection System
}
\examples{
# api keys can be obtained from your account on the http://quebio.ca/odk website
conn=list(user_id="JohnDoe",api_key="8fjasdf3h112aad")
my_forms<-qcbsmdc_list_forms(conn)
}
\seealso{
\code{\link{qcbsmdc_list_tables}} for obtaining list of tables for a particular form

\code{\link{qcbsmdc_get_tables}} for obtaining the content of the tables
}

