% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcbsmdc_list_tables.R
\name{qcbsmdc_list_tables}
\alias{qcbsmdc_list_tables}
\title{Get the list of tables for a given form and user on the QCBS Mobile Data Collection System}
\usage{
qcbsmdc_list_tables(conn, form_id)
}
\arguments{
\item{conn}{a connection list file containing user_id and api_key values}

\item{form_id}{the name or id of the form from which the table list is to be obtained}
}
\value{
character vector with list of table names
}
\description{
Get the list of tables for a given form and user on the QCBS Mobile Data Collection System
}
\examples{
## api keys can be obtained from your account on the http://quebio.ca/odk website
conn=list(user_id="JohnDoe",api_key='8fjasdf3h112aad') 
table_list<-qcbs_list_tables(conn,'MyForm1')
}
\seealso{
\code{\link{qcbsmdc_list_forms}} for obtaining forms

\code{\link{qcbsmdc_get_tables}} for obtaining the content of the tables
}

